requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2
idna==3.4
lxml==4.9.1
mysqlclient==2.1.1
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.2.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
text-unidecode==1.3
urllib3==1.26.12
